Refactor ArrayResize builtin
