+++ libs/mql-interpreter/src/builtins/impl/ArrayResize.ts
@@ -1,7 +1,15 @@
 import type { BuiltinFunction } from '../types';
-import { ArrayResize as arrayResize } from './helpers/arrayResize';
 
-export const ArrayResize: BuiltinFunction = (arr: any[], newSize: number) => {
-  arrayResize(arr, newSize);
+export const ArrayResize: BuiltinFunction = (
+  arr: any[],
+  newSize: number,
+  defaultValue?: any,
+) => {
+  if (newSize < 0) throw new Error('newSize must be >= 0');
+  if (arr.length > newSize) {
+    arr.length = newSize;
+  } else {
+    while (arr.length < newSize) arr.push(defaultValue);
+  }
   return arr.length;
 };
